requirements.txt
Pillow==9.3.0
pyocr==0.8.3
